requirements.txt
gradio
diffusers
transformers
torch
pillow
